+++ packages/liveManager/src/utils.ts
@@ -159,11 +159,19 @@ export function isFfmpegStartSegment(line: string) {
   return line.includes("Opening ") && line.includes("for writing");
 }
 
+export function isMesioStartSegment(line: string) {
+  return line.includes("Opening ") && line.includes("Opening FLV segment");
+}
+
 export function isFfmpegStart(line: string) {
   return (
     (line.includes("frame=") && line.includes("fps=")) ||
     (line.includes("speed=") && line.includes("time="))
   );
+}
+
+export function cleanTerminalText(text: string) {
+  return text.replace(/\x1B\[[0-9;]*[a-zA-Z]/g, "").replace(/[\x00-\x1F\x7F]/g, "");
 }
 
 export const formatTemplate = function template(string: string, ...args: any[]) {