+++ packages/shared/src/enum.ts
@@ -77,6 +77,7 @@ export const APP_DEFAULT_CONFIG: AppConfig = {
   ffprobePath: "",
   danmuFactoryPath: "",
   losslessCutPath: "",
+  mesioPath: "",
   /** 允许自定义可执行文件地址 */
   customExecPath: false,
   bilibiliUser: {},
@@ -246,6 +247,7 @@ export const APP_DEFAULT_CONFIG: AppConfig = {
     debugMode: false,
     qualityRetry: 0,
     videoFormat: "auto",
+    recorderType: "ffmpeg",
     useServerTimestamp: true,
     bilibili: {
       uid: undefined,