+++ packages/shared/src/recorder/index.ts
@@ -11,6 +11,7 @@ import { provider as providerForDouYin } from "@bililive-tools/douyin-recorder";
 import {
   createRecorderManager as createManager,
   setFFMPEGPath,
+  setMesioPath,
   utils,
 } from "@bililive-tools/manager";
 
@@ -110,8 +111,9 @@ export async function createRecorderManager(appConfig: AppConfig) {
   }
 
   const config = appConfig.getAll();
-  const { ffmpegPath } = getFfmpegPath();
+  const { ffmpegPath, mesioPath } = getFfmpegPath();
   setFFMPEGPath(ffmpegPath);
+  setMesioPath(mesioPath);
 
   const savePathRule = path.join(config?.recorder?.savePath, config?.recorder?.nameRule);
   const autoCheckInterval = config?.recorder?.checkInterval ?? 60;
@@ -135,7 +137,7 @@ export async function createRecorderManager(appConfig: AppConfig) {
     if (recorder.recordHandle) {
       const logFilePath = utils.replaceExtName(
         `${recorder.recordHandle.savePath}_${recorder.id}`,
-        ".ffmpeg.log",
+        ".recorder.log",
       );
       fs.appendFileSync(logFilePath, log.text + "\n");
       return;
@@ -317,8 +319,9 @@ export async function createRecorderManager(appConfig: AppConfig) {
   });
 
   appConfig.on("update", () => {
-    const { ffmpegPath } = getFfmpegPath();
+    const { ffmpegPath, mesioPath } = getFfmpegPath();
     setFFMPEGPath(ffmpegPath);
+    setMesioPath(mesioPath);
     updateRecorderManager(manager, appConfig);
   });