+++ packages/DouYinRecorder/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.6.0
+
+- 支持 `recorderType` 参数用于配置底层录制器，支持`ffmpeg | mesio`
+
 # 1.5.2
 
 - `resolveChannelInfoFromURL` 新增返回参数：`avatar`